+++ ChangeLog.md
@@ -1,4 +1,6 @@
 # Changelog for formark
 
 ## Unreleased changes
-
+* Removed custom parser
+* Using pandoc
+* using optparse-applicative